# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for ItsPlanned - A Collaborative Event Planning Application. This API provides endpoints for managing events, tasks, user profiles, and integrations with external services like Google Calendar.",
        "title": "ItsPlanned API",
        "contact": {
            "name": "ItsPlanned Support",
            "url": "https://github.com/vl4ddos/itsplanned",
            "email": "support@itsplanned.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/auth/oauth/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete the OAuth tokens for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "Delete OAuth tokens",
                "responses": {
                    "200": {
                        "description": "Token deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Token not found",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete token",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/oauth/save": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Save the OAuth tokens for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "Save OAuth tokens",
                "parameters": [
                    {
                        "description": "OAuth token details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SaveOAuthTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token saved successfully",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or unauthorized token save attempt",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to save token",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid input"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "user": {
                    "$ref": "#/definitions/api.UserResponse"
                }
            }
        },
...
        "api.SaveOAuthTokenRequest": {
            "type": "object",
            "required": [
                "access_token",
                "expiry"
            ],
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "ya29.a0AfB_byC..."
                },
                "expiry": {
                    "type": "string",
                    "example": "2024-03-16T15:04:05Z"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "1//04dK..."
                }
            }
        },
...
        "api.UserResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "bio": {
                    "type": "string",
                    "example": "Software developer and tech enthusiast"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-03-16T12:00:00Z"
                },
                "display_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-03-16T12:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}